refactor(component): update TempPrecipChart component and add Walter-Lieth chart functionality
- introduce new types for chart props and data handling in TempPrecipChart.type.ts.
- create WalterLiethChart component to visualize data using Walter-Lieth scales.
- add AridityLegend, SummaryStats, ModeToggle, and other components for enhanced chart interactivity and information display.
- implement custom hooks for managing chart data and state.
- add utility functions for building comparison data and computing chart summaries.
- add svg icons for chart mode toggling.
- establish a color utility for consistent chart styling.
