+++ src/components/TempPrecipChart/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useTempPrecipChart } from "./useTempPrecipChart";
