+++ src/components/TempPrecipChart/icons/IconAreaChart.tsx
@@ -1,0 +1,15 @@
+export function IconAreaChart() {
+  return (
+    <svg width="14" height="14" viewBox="0 0 14 14" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M1 11 L4 6 L7 8 L10 3 L13 5 L13 13 L1 13 Z" fill="currentColor" opacity="0.4" />
+      <path
+        d="M1 11 L4 6 L7 8 L10 3 L13 5"
+        stroke="currentColor"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        fill="none"
+      />
+    </svg>
+  );
+}
