+++ src/components/TempPrecipChart/icons/IconBarChart.tsx
@@ -1,0 +1,9 @@
+export function IconBarChart() {
+  return (
+    <svg width="14" height="14" viewBox="0 0 14 14" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <rect x="1" y="5" width="3" height="8" rx="0.5" fill="currentColor" />
+      <rect x="5.5" y="2" width="3" height="11" rx="0.5" fill="currentColor" />
+      <rect x="10" y="7" width="3" height="6" rx="0.5" fill="currentColor" />
+    </svg>
+  );
+}
