+++ src/components/TempPrecipChart/icons/index.ts
@@ -1,0 +1,2 @@
+export { IconAreaChart } from "./IconAreaChart";
+export { IconBarChart } from "./IconBarChart";
